third_party/WebKit/LayoutTests/fast/dom/shadow/shadowroot-clonenode-expected.txt

Calling ShadowRoot.cloneNode() should throw a DataCloneError exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.cloneNode() threw exception DataCloneError: Failed to execute 'cloneNode' on 'ShadowRoot': ShadowRoot nodes are not clonable..
PASS successfullyParsed is true

TEST COMPLETE
